Reassignable message parser
Instead of just enqueuing messages to be parsed in the Unity Update loop, when a message is received, it is sent to the System.Action<object> MessageParser function.

By default in UDP, this function simply enqueues the byte[] object passed in.

By default in TCP, this function will read from the NetworkStream object passed in until reading a complete message, before enqueueing.

By making this Action<object> public, it allows for other developers to make custom parsers that will parse/pre-parse messages in a thread as they arrive.